Fixed tests; fixed non-x86 builds.

+++ src/hotspot/cpu/x86/sharedRuntime_x86.cpp
@@ -124,7 +124,7 @@ jne    1b        \n\
   } else {
     jdouble (*addr)(jdouble, jdouble) = (double (*)(double, double))StubRoutines::fmod();
 
-    retval = (jfloat) (*addr)(x, y);
+    retval = (*addr)(x, y);
   }
   return retval;
 JRT_END
